+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 * st-switch: Added `position` input of type `StSwitchPosition` to change switch position. Defaults to left
 * st-radio: Added `showLabel` boolean input to show control's label. Defaults to true
 * st-select: Enable keyboard listener by default to open, close and move across menu options
-* st-select: Enable keyboard listener to search options (as native behaviour)
+* st-select: Enable keyboard listener to search options (as native MAC behaviour)
 * st-modal2: Added "fullWindow" input to apply full size to the modal
 * st-modal2: st-modal2 was renamed externally (not in code) as st-modal
 * st-modal2: Added functionality to close modal when clicking outside
@@ -52,6 +52,7 @@
 * st-color-picker: Removed component
 * st-two-list-selection: Remove component
 * st-dynamic-table: Removed component
+* st-select: Removed "keyBoardMove" input
 
 **Deprecated:**